tidied up collaboration template related tests + imports
