+++ test/functional-api/templates/collaboration/collaboration-template-testdata.ts
