+++ test/functional-api/templates/collaboration/collaboration-template.request.params.ts
@@ -1,52 +1,12 @@
 import { TestUser } from '@test/utils/token.helper';
+
+import { getGraphqlClient } from '@test/utils/graphqlClient';
+import { graphqlErrorWrapper } from '@test/utils/graphql.wrapper';
+import { TemplateType } from '@test/generated/alkemio-schema';
 import {
   lifecycleDefaultDefinition,
   templateDefaultInfo,
-} from './innovation-flow-template-testdata';
-import { getGraphqlClient } from '@test/utils/graphqlClient';
-import { graphqlErrorWrapper } from '@test/utils/graphql.wrapper';
-import { getSpaceData } from '../../journey/space/space.request.params';
-import { TemplateType } from '@test/generated/alkemio-schema';
-
-export const updateInnovationFlowState = async (
-  innovationFlowId: string,
-  selectedState: string,
-  userRole: TestUser = TestUser.GLOBAL_ADMIN
-) => {
-  const graphqlClient = getGraphqlClient();
-  const callback = (authToken: string | undefined) =>
-    graphqlClient.UpdateInnovationFlowSelectedState(
-      {
-        innovationFlowId,
-        selectedState,
-      },
-      {
-        authorization: `Bearer ${authToken}`,
-      }
-    );
-  return graphqlErrorWrapper(callback, userRole);
-};
-
-export const updateInnovationFlowStatesFromTemplate = async (
-  innovationFlowID: string,
-  innovationFlowTemplateID: string,
-  userRole: TestUser = TestUser.GLOBAL_ADMIN
-) => {
-  const graphqlClient = getGraphqlClient();
-  const callback = (authToken: string | undefined) =>
-    graphqlClient.UpdateInnovationFlowStatesFromTemplate(
-      {
-        input: {
-          innovationFlowID,
-          innovationFlowTemplateID,
-        },
-      },
-      {
-        authorization: `Bearer ${authToken}`,
-      }
-    );
-  return graphqlErrorWrapper(callback, userRole);
-};
+} from './collaboration-template-testdata';
 
 export const createInnovationFlowTemplate = async (
   templatesSetId: string,
@@ -119,29 +79,4 @@ export const deleteInnovationFlowTemplate = async (
       }
     );
   return graphqlErrorWrapper(callback, userRole);
-};
-
-export const getLifeCycleTemplateForSpaceByLifecycleTitle = async (
-  spaceId: string,
-  displayName: string
-) => {
-  const templatesPerSpace = await getSpaceData(spaceId);
-  const allTemplates =
-    templatesPerSpace?.data?.space?.library?.innovationFlowTemplates ?? [];
-
-  const filteredTemplate = allTemplates?.filter(item => {
-    return item.profile.displayName === displayName;
-  });
-
-  return filteredTemplate;
-};
-
-export const getInnovationFlowTemplatesCountForSpace = async (
-  spaceId: string
-) => {
-  const template = await getSpaceData(spaceId);
-  const spaceInnovationFlowTemplates =
-    template?.data?.space?.library?.innovationFlowTemplates.length;
-
-  return spaceInnovationFlowTemplates;
 };
